+++ static/app/components/draggableTabs/index.spec.tsx
