+++ static/app/components/draggableTabs/item.tsx
@@ -2,11 +2,11 @@ import {Item as _Item} from '@react-stately/collections';
 import type {ItemProps} from '@react-types/shared';
 import type {LocationDescriptor} from 'history';
 
-export interface TabListItemProps extends ItemProps<any> {
+export interface DraggableTabListItemProps extends ItemProps<any> {
   key: string | number;
   disabled?: boolean;
   hidden?: boolean;
   to?: LocationDescriptor;
 }
 
-export const Item = _Item as (props: TabListItemProps) => JSX.Element;
+export const Item = _Item as (props: DraggableTabListItemProps) => JSX.Element;
