--- static/app/components/tabs_dndTabs/utils.tsx
@@ -1,4 +1,0 @@
-import isPropValid from '@emotion/is-prop-valid';
-
-export const tabsShouldForwardProp = (prop: string) =>
-  typeof prop === 'string' && isPropValid(prop) && prop !== 'orientation';
